BackEnd/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
google-generativeai
sentence-transformers
numpy
json
psycopg2